fix: overflow in token supply

+++ test/recon/Setup.sol
@@ -49,7 +49,7 @@ abstract contract Setup is BaseSetup, MockStakingV1Deployer {
 
         (stakingV1, lqty, lusd) = deployMockStakingV1();
 
-        uint256 initialMintAmount = type(uint256).max;
+        uint256 initialMintAmount = type(uint88).max;
         lqty.mint(user, initialMintAmount);
         lqty.mint(user2, initialMintAmount);
         lusd.mint(user, initialMintAmount);
